I FINISHED THE PARSER
Everything parsing related is done and tested. Generics, loops, and
reference parameters/arguments are now all complete. FIxed memory leak
in infix operators, and added tests for every construct. Before moving
on to semantic analysis, I want to test the reconstruction of the AST,
add newlines for auto-formatting, and maybe add a node vtable function
for dumping. I believe the vtable needs to also have a way to support
node identification for semantic analysis, but that will take some
brainstorming.