Created test fixture for semantic analysis.
Also fixed parser test file name, which i feel really stupid about.
Broke up some header/source defs for test helpers. Also adjusted
signtature for some core seman functions so that I can grind out the
start of the real impl tomorrow!

The idea is that the signatures will not change for the analyzer again,
and instead any information will be stored within the context. This
should work for most implementations and should allow me to handle, for
example, functions with code paths that do not return a value. This may
evolve into me making an 'unreachable' statement as well.

Looking forward to this next chapter, though I know it will be difficult
and take a very long time!