Implemented reference counting semantics.
Memory model for semantic types is now completely different and fully
robust, supporting shared ownership properly.

Majority of the work in this commit is from include cleaning. I tried to
reduce the number of includes in headers to keep them lighter and force
source files to import actual things. There's still some overlap, but
it's much cleaner.

Want to try to stick with forward declaration when possible in the
future.

Ready to try doing enums again. Should be much simpler now that types
are reference counted and memory can't be corrupted as easily.