Added clang tidy support through dotfile and cmake.
Still need to clean up a few errors and tighten linting, but it looks a
lot better now. THis is a big step in the right direction.

Static analysis is crutial, and will be enforced for the remainder of
the projects work.

Will merge this into the main branch once I clean up static analysis and
add it to the workflow!