Array work, MSVC drop, and C23 migration (#6)
* Attempting to upgrade to c23 and finished implementation of array
parsers. Also addressed some lingering bugs in previous parsers and
analyzers. Going to move on to array type parsing next.

* Modernize codebase to use c23 features. MSVC may not work, which isnt
the end of the world.

Im happy to drop it from development.

* Remove support for MSVC.

* Transition from NODISCARD macro to [[nodiscard]].

Moving to removal of MAYBE_UNUSED next.

* Formatting and switch vtable to not used nodiscard since that only works
on clang.

* Removed usage of MAYBE_USED.

[[maybe_unused]] is used everywhere now, included in todo function
parameters. Ready to move on with work if checks pass.